Fix and move of test after rebase
